lastcommit

+++ frontend/src/components/WorkoutDetails.js
@@ -1,16 +1,22 @@
 import React from "react";
-import { BASE_URL } from "../utilities/helper";
+// import { BASE_URL } from "../utilities/helper";
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
-
+import { useAuthContext } from "../hooks/useAuthContext";
 // date fns
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
 
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutsContext();
+  const { user } = useAuthContext();
 
   const handleDelete = async () => {
-    const response = await fetch(`${BASE_URL}/api/workouts/` + workout._id, {
+    if (!user) {
+      return;
+    }
+
+    const response = await fetch("api/workouts/" + workout._id, {
       method: "DELETE",
+      Authorization: `Bearer ${user.token}`,
     });
 
     //  In here we have data of deleted document in database
